scripts/requirements-db.txt
# Dépendances pour la configuration de la base de données Supabase
psycopg2-binary>=2.9.3
python-dotenv>=0.19.0
